database/requirements.txt
psycopg2-binary==2.9.9
faker==22.0.0
